Fix: Case No Nota objects, saldo 0, etc.

+++ ABB.Application/EntriPembayaranBanks/Commands/UpdatePembayaranBankLihatCommand.cs
@@ -17,7 +17,7 @@ namespace ABB.Application.EntriPembayaranBanks.Commands
         public decimal? TotalBayar { get; set; }
         public string DebetKredit { get; set; }
         public decimal? TotalDlmRupiah { get; set; }
-        public int? Kurs { get; set; }
+        public decimal? Kurs { get; set; }
     }
 
     public class UpdatePembayaranBankLihatCommandHandler : IRequestHandler<UpdatePembayaranBankLihatCommand>
