+++ ABB.Application/EntriPembayaranKass/Commands/UpdatePembayaranKasCommand.cs
@@ -17,7 +17,7 @@ namespace ABB.Application.EntriPembayaranKass.Commands
         public string KodeMataUang { get; set; }
         public decimal? TotalBayar { get; set; }
         public string DebetKredit { get; set; }
-         public int? Kurs { get; set; }
+         public decimal? Kurs { get; set; }
         public decimal? TotalDlmRupiah { get; set; }
 
           public string KodeUserInput { get; set; }
