+++ ABB.Domain/Entities/EntriPembayaranBank.cs
@@ -18,7 +18,7 @@ namespace ABB.Domain.Entities
        public string DebetKredit { get; set; }
         public string UserBayar { get; set; }
         public decimal? TotalDlmRupiah { get; set; }
-        public int? Kurs { get; set; }
+        public decimal? Kurs { get; set; }
          public decimal? NilaiKurs { get; set; }
         //  public decimal? nilai_kurs { get; set; }
